Uploading the Jupyter Notebook as an html file
My Udacity workspace with GPU stopped working so had to reduce scope. Thus used the testing set as the training set and compared against the validation set which made it a 50% split training and testing. Accuracy was impacted by this reduced scope reaching 19% accuracy.

Before the technical issue before saving the accuracy reached was above 70% using only one epoch.

The saved model files with extension .pth can not be loaded to GitHub as max allowed size is 25 MB and files are 685 MB.